+++ src/relation/sql.rs
@@ -318,7 +318,9 @@ impl<'a> Visitor<'a, ast::Query> for FromRelationVisitor {
                     .collect(),
                 table_with_joins(reduce.input.as_ref().into(), vec![]),
                 None,
-                ast::GroupByExpr::Expressions(reduce.group_by.iter().map(ast::Expr::from).collect()),
+                ast::GroupByExpr::Expressions(
+                    reduce.group_by.iter().map(ast::Expr::from).collect(),
+                ),
                 vec![],
                 None,
             ),
@@ -451,7 +453,15 @@ impl<'a> Visitor<'a, ast::Query> for FromRelationVisitor {
             },
             joins: vec![],
         };
-        query(vec![], all(), from, None, ast::GroupByExpr::Expressions(vec![]), vec![], None)
+        query(
+            vec![],
+            all(),
+            from,
+            None,
+            ast::GroupByExpr::Expressions(vec![]),
+            vec![],
+            None,
+        )
     }
 }